+++ lib/fastlane/plugin/wpmreleasetoolkit/helper/metadata_download_helper.rb
@@ -4,23 +4,24 @@ require 'json'
 module Fastlane
   module Helper
     class MetadataDownloader
-      AUTO_RETRY_SLEEP_TIME = 20
-      MAX_AUTO_RETRY_ATTEMPTS = 30
-
       attr_reader :target_folder, :target_files
 
-      def initialize(target_folder, target_files, auto_retry)
+      def initialize(target_folder, target_files, auto_retry, auto_retry_sleep_time = 20)
         @target_folder = target_folder
         @target_files = target_files
         @auto_retry = auto_retry
+        @auto_retry_sleep_time = auto_retry_sleep_time
         @alternates = {}
         @auto_retry_attempt_counter = 0
       end
 
       # Downloads data from GlotPress, in JSON format
       def download(target_locale, glotpress_url, is_source)
-        uri = URI(glotpress_url)
-        response = Net::HTTP.get_response(uri)
+        downloader = GlotpressDownloader.new(
+          auto_retry: @auto_retry,
+          auto_retry_sleep_time: @auto_retry_sleep_time
+        )
+        response = downloader.download(glotpress_url)
         handle_glotpress_download(response: response, locale: target_locale, is_source: is_source)
       end
 
@@ -112,22 +113,6 @@ module Fastlane
           loc_data = JSON.parse(response.body) rescue loc_data = nil
           parse_data(locale, loc_data, is_source)
           reparse_alternates(target_locale, loc_data, is_source) unless @alternates.empty?
-        when '301'
-          # Follow the redirect
-          UI.message("Received 301 for `#{locale}`. Following redirect...")
-          download(locale, response.header['location'], is_source)
-        when '429'
-          # We got rate-limited, auto_retry or offer to try again with a prompt
-          if @auto_retry && @auto_retry_attempt_counter <= MAX_AUTO_RETRY_ATTEMPTS
-            UI.message("Received 429 for `#{locale}`. Auto retrying in #{AUTO_RETRY_SLEEP_TIME} seconds...")
-            sleep(AUTO_RETRY_SLEEP_TIME)
-            @auto_retry_attempt_counter += 1
-            download(locale, response.uri, is_source)
-          elsif UI.confirm("Retry downloading `#{locale}` after receiving 429 from the API?")
-            download(locale, response.uri, is_source)
-          else
-            UI.error("Abandoning `#{locale}` download as requested.")
-          end
         else
           message = "Received unexpected #{response.code} from request to URI #{response.uri}."
           UI.abort_with_message!(message) unless UI.confirm("#{message} Continue anyway?")